+++ attribution/src/main/java/com/affise/attribution/events/custom/UserCustomSubscriptionEvent.kt
@@ -1,5 +1,6 @@
 package com.affise.attribution.events.custom
 
+import com.affise.attribution.events.Event
 import com.affise.attribution.events.subscription.BaseSubscriptionEvent
 import org.json.JSONObject
 
@@ -14,5 +15,17 @@ class UserCustomSubscriptionEvent(
     override val type: String,
     override val subtype: String,
     data: JSONObject,
-    userData: String? = null
-) : BaseSubscriptionEvent(data, userData)
+    userData: String? = null,
+    private val category: String? = null,
+) : BaseSubscriptionEvent(data, userData) {
+
+    override fun getCategory(): String = category ?: super.getCategory()
+
+    /**
+     * For internal use only
+     */
+    fun internalAddRawParameters(parameters: Map<String, Any>): Event {
+        addRawParameters(parameters)
+        return this
+    }
+}