+++ attribution/src/main/java/com/affise/attribution/modules/AffiseModuleApiWrapper.kt
@@ -1,0 +1,13 @@
+package com.affise.attribution.modules
+
+import com.affise.attribution.Affise
+
+abstract class AffiseModuleApiWrapper<API : AffiseModuleApi>(
+    module: AffiseModules
+) {
+    private var api: API? = null
+
+    protected val moduleApi: API? = api ?: Affise.Module.api<API>(module).also {
+        api = it
+    }
+}