+++ attribution/src/main/java/com/affise/attribution/modules/AffiseModules.kt
@@ -2,16 +2,19 @@ package com.affise.attribution.modules
 
 
 enum class AffiseModules(val module: String) {
-    Advertising("com.affise.attribution.module.advertising.AdvertisingModule"),
-    AndroidId("com.affise.attribution.module.androidid.AndroidIdModule"),
-    Link("com.affise.attribution.module.link.LinkModule"),
-    Network("com.affise.attribution.module.network.NetworkModule"),
-    Phone("com.affise.attribution.module.phone.PhoneModule"),
-    Status("com.affise.attribution.module.status.StatusModule"),
-    Subscription("com.affise.attribution.module.subscription.SubscriptionModule"),
-    RuStore("com.affise.attribution.module.rustore.RuStoreModule"),
-    Huawei("com.affise.attribution.module.huawei.HuaweiModule"),
-    Meta("com.affise.attribution.module.meta.MetaModule");
+    Advertising("Advertising"),
+    AndroidId("AndroidId"),
+    AppsFlyer("AppsFlyer"),
+    Link("Link"),
+    Network("Network"),
+    Phone("Phone"),
+    Status("Status"),
+    Subscription("Subscription"),
+    RuStore("RuStore"),
+    Huawei("Huawei"),
+    Meta("Meta");
+
+    internal val className: String = "${AffiseModules.MODULE_PREFIX}.${this.module.lowercase()}.${this.module}Module"
 
     companion object {
         @JvmStatic
@@ -19,6 +22,8 @@ enum class AffiseModules(val module: String) {
             name ?:  return null
             return values().firstOrNull { it.module.contains(name, true) }
         }
+
+        private const val MODULE_PREFIX = "com.affise.attribution.module"
     }
 }