+++ attribution/src/main/java/com/affise/attribution/modules/appsflyer/AffiseAppsFlyerApi.kt
@@ -1,0 +1,7 @@
+package com.affise.attribution.modules.appsflyer
+
+import com.affise.attribution.modules.AffiseModuleApi
+
+interface AffiseAppsFlyerApi : AffiseModuleApi {
+    fun logEvent(eventName: String, eventValues: Map<String, Any>)
+}