+++ attribution/src/main/java/com/affise/attribution/modules/link/AffiseLink.kt
@@ -1,23 +1,16 @@
 package com.affise.attribution.modules.link
 
 import com.affise.attribution.Affise
+import com.affise.attribution.modules.AffiseModuleApiWrapper
 import com.affise.attribution.modules.AffiseModules
 
-object AffiseLink {
-
-    private val api: AffiseLinkApi?
-        get() = module ?: Affise.Module.api<AffiseLinkApi>(AffiseModules.Link).also {
-            module = it
-        }
-
-    private var module: AffiseLinkApi? = null
-
+object AffiseLink : AffiseModuleApiWrapper<AffiseLinkApi>(AffiseModules.Link) {
     /**
      * Module Link url Resolve
      */
     @JvmStatic
     fun linkResolve(url: String, callback: AffiseLinkCallback) {
-        api?.linkResolve(url, callback) ?: callback.handle("")
+        moduleApi?.linkResolve(url, callback) ?: callback.handle("")
     }
 }