+++ internal/src/main/java/com/affise/attribution/internal/events/PredefinedEvent.kt
@@ -2,6 +2,7 @@ package com.affise.attribution.internal.events
 
 import com.affise.attribution.events.Event
 import com.affise.attribution.events.custom.UserCustomEvent
+import com.affise.attribution.internal.utils.getCategory
 import com.affise.attribution.internal.utils.getTimeStamp
 import com.affise.attribution.internal.utils.getUserData
 import com.affise.attribution.utils.timestamp
@@ -12,11 +13,13 @@ internal object PredefinedEvent {
         name ?: return null
         val timeStamp = map.getTimeStamp() ?: timestamp()
         val userData = map.getUserData()
+        val category =  map.getCategory()
 
         return UserCustomEvent(
             eventName = name,
             userData = userData,
-            timeStampMillis = timeStamp
+            timeStampMillis = timeStamp,
+            category = category,
         )
     }
 }