+++ module-appsflyer/src/main/java/com/affise/attribution/module/appsflyer/usecase/AppsFlyerUseCase.kt
@@ -1,0 +1,6 @@
+package com.affise.attribution.module.appsflyer.usecase
+
+internal interface AppsFlyerUseCase {
+
+    fun logEvent(eventName: String, eventValues: Map<String, Any>)
+}